Updating shop
