+++ client/my-app/src/App.js
@@ -8,6 +8,8 @@ import Contact from "./Contact";
 import Catalog from "./Catalog";
 import './App.css';
 import Cursor from "./Cursor";
+import ShopDisplay from "./ShopDisplay";
+import newlogo1 from './images/newlogo1.png'
 
 
 function App() {
@@ -21,6 +23,9 @@ function App() {
 	  	<Route path='/catalog' element={<Catalog/>} />
         <Route path='/aboutUs' element={<AboutUs/>} />
         <Route path='/aboutUsInfo' element={<AboutUsInfo/>} />
+		<Route path="/shop/1" element = {<ShopDisplay name = "bubblegumblush" id="1" src = {newlogo1}/>} />
+		<Route path="/shop/2" element = {<ShopDisplay name = "fuchsiafusion" id="2" src = {newlogo1}/>} />
+		<Route path="/shop/3" element = {<ShopDisplay name = "strawberrysorbet" id="3" src = {newlogo1}/>} />
 	  </Routes>
       <Cursor />
     </div>
